CanopyRadiation

Types

CanopyRadiation.Canopy4RT — Type

struct Canopy4RT

A canopy struct for the radiation transfer module

Fields

nLayer

Number of canopy layers

LAI

Leaf Area Index

Ω

Clumping factor

clump_a

Structure factor a

clump_b

Structure factor b

leaf_width

Leaf width

hc

Vegetation height

LIDFa

Leaf Inclination

LIDFb

Variation in leaf inclination

hot

HotSpot parameter (still need to check!)

height

Canopy height [m]

z0m

Canopy roughness [m]

z0h

Tree roughtnes [m]

d

Canopy displacement height [m]

Cd

m/sqrt(s) turbulent transfer coefficient

litab

List of mean inclination angles [°]

litab_bnd

List of inclination angle boundaries [°]

lazitab

List of mean azimuth angles [°]

cos_ttlo

Cosine of lazitab

cos_philo

Cosine of lazitab - psi (relative azimuth angle), update with time

cos_ttli

Cosine of litab

sin_ttli

Sine of litab

vol_scatt

Container for volome scatter function

lidf

Inclination angles weight distribution

xl

List of level location (level = layer + 1)

xl_e

Exp(xl)

dx

1/nLayer

nAzi

Number of azimuth angles

nIncl

Number of inclination angles

CanopyRadiation.CanopyOpticals — Type

struct CanopyOpticals{FT}

A struct for canopy optical properties using Array

Fields

nAzi

Number of azimuth angles

nIncl

Number of inclination agles

nLayer

Number of canopy layers

nWL

Number of wave lengths

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

?

sob

Weight of specular2directional backscatter coefficient

sof

Weight of specular2directional forward coefficient

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

absfo

abs(fo)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

CanopyRadiation.CanopyRads — Type

mutable struct CanopyRads{FT}

A struct for canopy radiation information

Fields

nAzi

Number of azimuth angles

nIncl

Number of inclination agles

nLayer

Number of canopy layers

nLevel

Number of canopy levels

nWL

Number of wave lengths

nWLF

Number of wave lengths for SIF

intEout

Integrated TOC outgoing flux [W m⁻²]

incomingPAR

Incident spectrally integrated total PAR [mol m⁻² s⁻¹]

incomingPAR_direct

Incident spectrally integrated direct PAR [mol m⁻² s⁻¹]

incomingPAR_diffuse

Incident spectrally integrated diffuse PAR [mol m⁻² s⁻¹]

RnSoil_diffuse

Net radiation of shaded soil [W m⁻²]

RnSoil_direct

Net Short-wave radiation of sunlit soil [W m⁻²]

RnSoil

Net Short-wave radiation of soil (shaded + sunlit) [W m⁻²]

RnSoilLW

Net long-wave radiation of soil (shaded + sunlit) [W m⁻²]

absPAR_shade

Net PAR of shaded leaves [mol m⁻² s⁻¹]

absPAR_shadeCab

Net PAR by Cab+Car of shaded leaves [moles m⁻² s⁻¹]

intNetSW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetSW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

intNetLW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetLW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

T_sun

Leaf temperature (sunlit) [K]

T_shade

Leaf temperature (shaded) [K]

ϕ_shade

Fluorescence yield for shaded leaves

H_shade

Sensible Heat flux H of shaded leaves [W m⁻²]

LE_shade

Latent Heat flux LE of shaded leaves [W m⁻²]

NPQ_shade

NPQ of shaded leaves

GPP_shade

GPP of shaded leaves [μmol m⁻² s⁻¹]

gs_shade

gs of shaded leaves [mol m⁻² s⁻¹]

ψl_shade

Leaf water potential of shaded leaves [MPa]

Cc_shade

Cc of shaded leaves [µmol/mol]

Pi_shade

internal CO₂ concentration of shaded leaves [µmol/mol]

Lo

Short-wave TOC outgoing radiance in observation direction [mW m⁻² nm⁻¹ sr⁻¹]

Eout

Short-wave TOC outgoing radiation [mW m⁻² nm⁻¹]

alb_obs

Short-wave Albedo in viewing direction

alb_direct

Short-wave Albedo for direct incoming radiation

alb_diffuse

Short-wave Albedo for diffuse incoming radiation

E_up

Upwelling diffuse short-wave radiation within canopy [mW m⁻² nm⁻¹]

E_down

Downwelling diffuse short-wave radiation within canopy [mW m⁻² nm⁻¹]

netSW_sunlit

Net absorbed direct radiation in each layer [mW m⁻² nm⁻¹]

netSW_shade

net absorbed diffuse radiation in each layer [mW m⁻² nm⁻¹]

absPAR_sun

net PAR of sunlit leaves [mol m⁻² s⁻¹]

absPAR_sunCab

net PAR by Cab+Car of sunlit leaves [mol m⁻² s⁻¹]

T_sun3D

Leaf temperature (sunlit) [K]

ϕ_sun

Fluorescence yield for sunlit leaves

H_sun

Sensible Heat flux H of sunlit leaves [W m⁻²]

LE_sun

Latent Heat flux LE of sunlit leaves [W m⁻²]

NPQ_sun

NPQ of sunlit leaves

GPP_sun

GPP of sunlit leaves [μmol m⁻² s⁻¹]

gs_sun

gs of sunlit leaves [mol m⁻² s⁻¹]

ψl_sun

Leaf water potential of sunlit leaves [MPa]

Cc_sun

Cc of sunlit leaves [µmol/mol]

Pi_sun

Internal CO₂ concentration of sunlit leaves [µmol/mol]

SIF_hemi

Hemispheric total outgoing SIF flux [mW m⁻² nm⁻¹])

SIF_obs

Observer-direction outgoing SIF radiance (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_sunlit

Observer-direction outgoing SIF radiance, sunlit leaves (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_shaded

Observer-direction outgoing SIF radiance, shaded leaves (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_scattered

Observer-direction outgoing SIF radiance, scattered (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_soil

Observer-direction outgoing SIF radiance, soil-reflected (mW m⁻² nm⁻¹ sr⁻¹))

SIF_sum

Total SIF sum of layer sources [mW m⁻² nm⁻¹])

CanopyRadiation.IncomingRadiation — Type

struct IncomingRadiation{FT}

Incoming radiation information.

Fields

E_direct

Direct incoming radiation [mW m⁻² nm⁻¹]

E_diffuse

Diffuse incoming radiation [mW m⁻² nm⁻¹]

CanopyRadiation.LeafBios — Type

struct LeafBios{FT}

A struct of leaf biological parameters

Fields

nWL

Number of wave length

nWLE

Number of wave length for excitation

nWLF

Number of wave length for SIF

N

Leaf structure parameter

Cab

Chlorophyll a+b content [µg cm⁻²]

Car

Carotenoid content [µg cm⁻²]

Ant

Anthocynanin content [µg cm⁻²]

Cs

Senescent material fraction

Cw

Equivalent water thickness [cm]

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin) (-)

ρ_LW

Broadband thermal reflectance (-)

τ_LW

Broadband thermal transmission (-)

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

Shortwave leaf reflectance

τ_SW

Shortwave leaf transmission

kChlrel

Relative absorbtion by Chlorophyll+Car

kChlrel_old

Relative absorbtion by Chlorophyll

Mb

Fluorescence excitation matrix backwards

Mf

Fluorescence excitation matrix forwards

ndub

Doubling adding layers

CanopyRadiation.LeafOpticals — Type

mutable struct LeafOpticals{FT}

Struct for leaf optical properties using Array

Fields

nr

Km

Kab

Kant

Kcar

Kw

KBrown

phi

KcaV

KcaZ

lambda

CanopyRadiation.SoilOpticals — Type

mutable struct SoilOpticals{FT}

A struct of soil optical parameters

Fields

albedo_SW

Shortwave albedo

albedo_SW_SIF

Shortwave albedo for SIF

emsvty_SW

Shortwave Emissivity

albedo_LW

Longwave albedo

soil_skinT

Soil surface temperature [K]

CanopyRadiation.SolarAngles — Type

struct SolarAngles{FT}

Struct for observation and solar angles

Fields

tts

Solar Zenith Angle [degree]

tto

Viewing Zenith Angle in [degree]

psi

relative azimuth in [degree]

CanopyRadiation.WaveLengths — Type

struct WaveLengths{FT}

Struct for pre-set wave length parameters.

Fields

minwlPAR

Minimal WL for PAR [nm]

maxwlPAR

Maximal WL for PAR [nm]

minwle

Minimal WL for SIF excitation [nm]

maxwle

Maximal WL for SIF excitation [nm]

minwlf

Minimal WL for SIF emission/fluorescence [nm]

maxwlf

Maximal WL for SIF emission/fluorescence [nm]

sWL

Standard wave length [nm]

dWL

Differential wavelength

optis

Leaf optical parameter set

WL

Wave length [nm]

iWLE

Index of WLE in WL

iWLF

Index of WLF in WL

iPAR

index of wlPAR in WL

WLE

excitation wave length [nm]

WLF

Fluorescence wave length [nm]

WL_iPAR

Wave length for PAR

dWL_iPAR

Differential wave length for PAR

dWL_iWLE

Differential wave length for iWLE

nPAR

Length of WL_iPAR

nWL

Length of WL

nWLE

length of WLE

nWLF

length of WLF

CanopyRadiation.create_canopy_opticals — Function

createcanopyopticals(FT, nWL::Int, nLayer::Int, nAzi::Int, nIncl::Int)

Create a canopy optical properties struct, given

FT Floating number type

nWL Number of wave length

nLayer Number of canopy layers

nAzi Number of arimuth angles

AIncl Number of inclination angles

CanopyRadiation.create_incoming_radiation — Function

create_incoming_radiation(sWL:Array, file)

Create an AbstractIncomingRadiation struct, given

FType Floating number type

sWL Standard wave length

file Input file name

CanopyRadiation.create_leaf_bios — Function

create_leaf_bios(FT, nWL::Int, nWLE::Int, nWLF::Int)

Create a leaf biological parameters struct, given

FType Floating number type

nWL Number of wave length

nWLE Number of excitation wave length

nWLF Number of fluorescence wave length

Returns a LeafBios type struct.

CanopyRadiation.create_leaf_opticals — Function

create_leaf_opticals(sWL::Array, file::String)

Create an AbstractLeafOptiPara struct, given

sWL Standard wave length

file Input file name

Utils

CanopyRadiation.calctav — Function

calctav(α::FT, nr::FT) where {FT<:AbstractFloat}

Computes transmission of isotropic radiation across an interface between two dielectrics (Stern F. (1964), Allen W.A. (1973)). From calctav.m in PROSPECT-D

α angle of incidence

nr Index of refraction

CanopyRadiation.e2phot — Function

e2phot(λ::Array{FT}, E::Array{FT}) where {FT<:AbstractFloat}

Calculates the number of moles of photons, given

λ An array of wave length in [nm], converted to [m] by _FAC

E Joules of energy

CanopyRadiation.fast∫ — Function

fast∫(f::Vector{FT}, dx::Vector{FT}) where {FT<:AbstractFloat}

A fast way of integrating functions, given

f f(x) for each x

dx Delta x for each x

Missing docstring.

Missing docstring for volscatt. Check Documenter's build log for details.

Big-leaf model

CanopyRadiation.big_leaf_partition — Function

big_leaf_partition(lai::FT, zenith::FT, r_all::FT) where {FT <:AbstractFloat}

Partition the big-leaf canopy into sunlit and shaded layers, given

partition Container for partition

zenith Zenith angle in degree

r_all Total radiation in [W m⁻²]

Layered model

CanopyRadiation.canopy_fluxes! — Function

canopy_fluxes!(can::Canopy4RT{FT}, can_opt::CanopyOpticals{FT}, can_rad::CanopyRads{FT}, in_rad::IncomingRadiation{FT}, soil_opt::SoilOpticals{FT}, leaf_array::Array{LeafBios{FT},1}, wl_set::WaveLengths{FT}, rt_con::RTContainer) where {FT<:AbstractFloat}

Computes a variety of integrated fluxes from the spectrally resolved computations in the short-wave Canopy RT (e.g. absorbed soil radiation, absorbed direct and diffuse PAR by layer (and angles for direct), net direct and diffuse energy balance per layer), given

can A Canopy4RT struct

can_opt A CanopyOpticals struct

can_rad A CanopyRads struct

in_rad An IncomingRadiation struct

soil_opt A SoilOpticals type struct for soil optical properties

leaf_array An array of LeafBios type struct (i.e. leaf optical properties can change with canopy height)

wl_set An WaveLengths type struct

rt_con RTContainer type container

CanopyRadiation.canopy_geometry! — Function

canopy_geometry!(can::Canopy4RT{FT}, angles::SolarAngles{FT}, can_opt::CanopyOpticals{FT}, rt_con::RTContainer{FT}) where {FT<:AbstractFloat}

Computes canopy optical properties (extinction coefficients for direct and diffuse light) based on the SAIL model. Most important input parameters are leaf inclination and azimuth distribution functions and sun-sensor geometry . Canopy clumping Ω is implemented as in Pinty et al (2015), given

can Canopy4RT type struct, providing canopy structure information

angles SolarAngles type struct, defining sun-sensor geometry

can_opt CanopyOpticals type, where optical properties is stored

rt_con RTContainer type container

CanopyRadiation.canopy_matrices! — Function

canopy_matrices!(leaf_array::Array{LeafBios{FT},1}, can_opt::CanopyOpticals{FT}) where {FT<:AbstractFloat}

Compute scattering coefficient matrices for direct and diffuse light given geometry dependent overall extinction coefficients and pigment dependent leaf reflectance and transmission (computed via fluspect). This function has to be called before simulate_short_wave! can be used.

leaf_array An array of LeafBios type struct (i.e. leaf optical properties can change with canopy height)

can_opt An CanopyOpticals type struct, will be updated within this function call.

CanopyRadiation.diffusive_S — Function

diffusive_S(τ_dd::Array{FT}, ρ_dd::Array{FT}, S⁻::Array{FT}, S⁺::Array{FT}, boundary_top::Array{FT}, boundary_bottom::Array{FT}, rsoil::Array{FT}) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport (used for thermal and SIF) given:

τ_dd A 2D Array with layer reflectances

ρ_dd A 2D Array with layer transmissions

S⁻ A 2D Array with layer source terms in the downwelling direction

S⁺ A 2D Array with layer source terms in the upwelling direction

boundary_top A 1D array with downwelling radiation at the top (top of canopy)

boundary_bottom A 1D array with upwnwelling radiation at the bottom (soil)

rsoil A 1D array with soil reflectance

CanopyRadiation.fluspect! — Function

fluspect!(leaf::LeafBios{FT}, wl_set::WaveLengths{FT}) where {FT<:AbstractFloat}

Computes leaf optical properties (reflectance and transittance) based on pigment concentrations. Also computes Fluorescence excitation matrices. Mostly based on PROSPECT-D for leaf reflectance/transmission and FluSpec for fluorescence.

leaf LeafBios type struct (includes pigment concentrations, water content, leaf structure)

wl_set An WaveLengths type struct, which defines fluoresence excitation and emission wavelengths

CanopyRadiation.initialize_rt_module — Function

initialize_rt_module(FT; nLayer::Int=20, LAI::FT=FT(3.0))

Initialize the RT module so as to interface with Plant, given

nLayer Number of canopy layers

LAI Leaf area index

Note it here that the structcanopy is used in CanopyRT module. This function initializes `canopyrt,canOptrt,canRadrt,arrayOfLeaves` as local variables rather than global variables as in CanopyRT module. See CanopyRT module for further operations on these variables.

CanopyRadiation.short_wave! — Function

short_wave!(can::Canopy4RT{FT}, can_opt::CanopyOpticals, can_rad::CanopyRads{FT}, in_rad::IncomingRadiation{FT}, soil_opt::SoilOpticals{FT}, rt_con::RTContainer{FT}) where {FT<:AbstractFloat}

Simulate the short wave radiation through the canopy, given

can Canopy4RT type struct of canopy information

can_opt CanopyOpticals struct of optical layer properties

can_rad A CanopyRads struct

in_rad An IncomingRadiation struct

soil_opt A SoilOpticals type struct for soil optical properties

rt_con RTContainer type container

CanopyRadiation.SIF_fluxes! — Function

SIF_fluxes!(leaf_array::Array{LeafBios{FT},1}, can_opt::CanopyOpticals{FT}, can_rad::CanopyRads{FT}, can::Canopy4RT{FT}, soil_opt::SoilOpticals{FT}, wl_set::WaveLengths{FT}) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport for SIF radiation (calls compute_diffusive_S internally). Layer reflectance and transmission is computed from SW optical properties, layer sources from absorbed light and SIF efficiencies. Boundary conditions are zero SIF incoming from atmosphere or soil.

leaf_array An array of LeafBios type struct (i.e. leaf optical properties can change with canopy height)

can_opt A CanopyOpticals struct for providing optical layer properties

can_rad A CanopyRads struct

can A Canopy4RT type struct for providing LAI and nLayer and clumping

soil_opt A SoilOpticals type struct for soil optical properties

wl_set An WaveLengths type struct

CanopyRadiation.thermal_fluxes! — Function

thermal_fluxes!(leaf_array::Array{LeafBios{FT},1}, can_opt::CanopyOpticals{FT}, can_rad::CanopyRads{FT}, can::Canopy4RT{FT}, soil_opt::SoilOpticals{FT}, incLW::Array{FT}, wl_set::WaveLengths{FT}) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport for thermal radiation (calls compute_diffusive_S internally). Layer reflectance and transmission is computed from LW optical properties, layer sources from temperature and Planck law, boundary conditions from the atmosphere and soil emissivity and temperature. Currently only uses Stefan Boltzmann law to compute spectrally integrated LW but can be easily adjusted to be spectrally resolved.

leaf_array An array of LeafBios type struct (i.e. leaf optical properties can change with canopy height)

can_opt A CanopyOpticals struct for providing optical layer properties

can_rad A CanopyRads struct

can A Canopy4RT type struct for providing LAI and nLayer and clumping

sO A SoilOpticals type struct for soil optical properties

incLW A 1D array with incoming long-wave radiation

wl_set An WaveLengths type struct